refactor: reducer separated + enum action types
